Version 3.5.5
git-svn-id: https://plugins.svn.wordpress.org/woocommerce-multilingual/trunk@1124368 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ inc/multi-currency-support.class.php
@@ -8,10 +8,11 @@ class WCML_Multi_Currency_Support{
     private $client_currency;
     private $exchange_rates = array();
     
-    
     function __construct(){
         
         add_action('init', array($this, 'init'), 5);
+        $this->install();
+
         $this->init_currencies();
 
         if(is_ajax()){        
@@ -29,12 +30,11 @@ class WCML_Multi_Currency_Support{
             add_action('wp_ajax_wcml_update_currency_lang', array($this,'update_currency_lang'));
             add_action('wp_ajax_wcml_update_default_currency', array($this,'update_default_currency'));
             
-            
         }
         
         if(is_admin()){
             add_action('admin_footer', array($this, 'currency_options_wc_integration'));            
-            add_action('woocommerce_settings_save_general', array($this, 'currency_options_wc_integration_save_hook'));            
+            add_action('woocommerce_settings_save_general', array($this, 'currency_options_wc_integration_save_hook'));
         }
         
         add_action( 'wp_enqueue_scripts', array( $this, 'register_styles' ) );
@@ -42,6 +42,8 @@ class WCML_Multi_Currency_Support{
 
         add_filter( 'woocommerce_cart_contents_total', array( $this, 'filter_woocommerce_cart_contents_total'), 100 );
         add_filter( 'woocommerce_cart_subtotal', array( $this, 'filter_woocommerce_cart_subtotal'), 100, 3 );
+
+        add_action( 'update_option_woocommerce_currency', array( $this, 'set_default_currencies_languages' ), 10, 2 );
     }
     
     function _load_filters(){
@@ -51,15 +53,15 @@ class WCML_Multi_Currency_Support{
             $load = true;
         }else{
             if(is_ajax() && $this->get_client_currency() != get_option('woocommerce_currency')){
-                if(isset($_REQUEST['action'])){
-                    $ajax_actions = apply_filters('wcml_multi_currency_is_ajax', array('woocommerce_get_refreshed_fragments', 'woocommerce_update_order_review', 'woocommerce-checkout', 'woocommerce_checkout', 'woocommerce_add_to_cart')); 
-                    if(version_compare($GLOBALS['woocommerce']->version, '2.1', '>=')){
-                        $ajax_actions[] = 'woocommerce_update_shipping_method';
-                    }
-                    if(in_array($_REQUEST['action'], $ajax_actions)){
-                        $load = true;
-                    }
+
+                $ajax_actions = apply_filters('wcml_multi_currency_is_ajax', array('woocommerce_get_refreshed_fragments', 'woocommerce_update_order_review', 'woocommerce-checkout', 'woocommerce_checkout', 'woocommerce_add_to_cart'));
+                if(version_compare($GLOBALS['woocommerce']->version, '2.1', '>=')){
+                    $ajax_actions[] = 'woocommerce_update_shipping_method';
                 }
+                if( ( isset( $_POST['action'] ) && in_array( $_POST['action'], $ajax_actions ) ) || (  isset( $_GET['action'] ) && in_array( $_GET['action'], $ajax_actions ) ) ){
+                    $load = true;
+                }
+
             }
         }
         
@@ -67,7 +69,6 @@ class WCML_Multi_Currency_Support{
     }
     
     function init(){
-        
         
         if($this->_load_filters()){    
             
@@ -114,8 +115,22 @@ class WCML_Multi_Currency_Support{
 
         if(!is_admin()) $this->load_inline_js();
         
-    }    
-    
+    }
+
+    function install(){
+        global $woocommerce_wpml;
+
+        if(empty($woocommerce_wpml->settings['multi_currency']['set_up'])){
+            $woocommerce_wpml->settings['multi_currency']['set_up'] = 1;
+            $woocommerce_wpml->update_settings();
+
+            $this->set_default_currencies_languages();
+        }
+
+        return;
+
+    }
+
     function init_currencies(){
         global $woocommerce_wpml, $sitepress;
         $this->currencies =& $woocommerce_wpml->settings['currency_options'];  // ref
@@ -197,9 +212,40 @@ class WCML_Multi_Currency_Support{
     function get_currency_codes(){
         return $this->currency_codes;
     }
+
+    function set_default_currencies_languages( $old_value = false, $new_value = false ){
+        global $woocommerce_wpml,$sitepress;
+
+        $settings = $woocommerce_wpml->get_settings();
+
+        $wc_currency = $new_value ? $new_value : get_option('woocommerce_currency');
+
+        $active_languages = $sitepress->get_active_languages();
+        foreach ($this->get_currency_codes() as $code) {
+            foreach($active_languages as $language){
+                if(!isset($settings['currency_options'][$code]['languages'][$language['code']])){
+                    $settings['currency_options'][$code]['languages'][$language['code']] = 1;
+                }
+            }
+        }
+
+        foreach($active_languages as $language){
+            if(!isset($settings['default_currencies'][$language['code']])){
+                $settings['default_currencies'][$language['code']] = false;
+            }
+
+            if(!isset($settings['currency_options'][$wc_currency]['languages'][$language['code']])){
+                $settings['currency_options'][$wc_currency]['languages'][$language['code']] = 1;
+            }
+        }
+
+        $woocommerce_wpml->update_settings($settings);
+
+    }
     
     function add_currency(){
-        if(!wp_verify_nonce($_REQUEST['wcml_nonce'], 'wcml_new_currency')){
+        $nonce = filter_input( INPUT_POST, 'wcml_nonce', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+        if(!$nonce || !wp_verify_nonce($nonce, 'wcml_new_currency')){
             die('Invalid nonce');
         }
 
@@ -210,7 +256,7 @@ class WCML_Multi_Currency_Support{
         
         if(!empty($_POST['currency_code'])){
             
-            $currency_code = $_POST['currency_code'];
+            $currency_code = filter_input( INPUT_POST, 'currency_code', FILTER_SANITIZE_FULL_SPECIAL_CHARS);
             
             $active_languages = $sitepress->get_active_languages();
             $return['languages'] ='';
@@ -219,7 +265,7 @@ class WCML_Multi_Currency_Support{
                     $settings['currency_options'][$currency_code]['languages'][$language['code']] = 1;
                 }
             }
-            $settings['currency_options'][$currency_code]['rate'] = (double) $_POST['currency_value'];
+            $settings['currency_options'][$currency_code]['rate'] = (double) filter_input( INPUT_POST, 'currency_value', FILTER_SANITIZE_NUMBER_FLOAT );
             $settings['currency_options'][$currency_code]['updated'] = date('Y-m-d H:i:s');        
 
             $wc_currency = get_option('woocommerce_currency'); 
@@ -243,9 +289,7 @@ class WCML_Multi_Currency_Support{
             include WCML_PLUGIN_PATH . '/menu/sub/custom-currency-options.php'; 
             $return['currency_options'] = ob_get_contents();
             ob_end_clean();
-            
-            
-            
+
         }
         
         echo json_encode($return);
@@ -253,9 +297,14 @@ class WCML_Multi_Currency_Support{
     }    
     
     function save_currency(){
+        $nonce = filter_input( INPUT_POST, 'wcml_nonce', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+        if(!$nonce || !wp_verify_nonce($nonce, 'save_currency')){
+            die('Invalid nonce');
+        }
+
         global $woocommerce_wpml;
         
-        $currency_code = $_POST['currency'];
+        $currency_code = filter_input( INPUT_POST, 'currency', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
         $options = $_POST['currency_options'][$currency_code];
         
         $changed = false;
@@ -299,7 +348,8 @@ class WCML_Multi_Currency_Support{
     }
     
     function delete_currency(){
-        if(!wp_verify_nonce($_REQUEST['wcml_nonce'], 'wcml_delete_currency')){
+        $nonce = filter_input( INPUT_POST, 'wcml_nonce', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+        if(!$nonce || !wp_verify_nonce($nonce, 'wcml_delete_currency')){
             die('Invalid nonce');
         }
         global $woocommerce_wpml;
@@ -319,9 +369,11 @@ class WCML_Multi_Currency_Support{
     }
     
     function currencies_list(){
-        if(!wp_verify_nonce($_REQUEST['wcml_nonce'], 'wcml_currencies_list')){
+        $nonce = filter_input( INPUT_POST, 'wcml_nonce', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+        if(!$nonce || !wp_verify_nonce($nonce, 'wcml_currencies_list')){
             die('Invalid nonce');
         }
+
         global $woocommerce_wpml;
         $wc_currencies = get_woocommerce_currencies();
         $wc_currency = get_option('woocommerce_currency');
@@ -341,7 +393,8 @@ class WCML_Multi_Currency_Support{
     }
     
     function update_currency_lang(){
-        if(!wp_verify_nonce($_REQUEST['wcml_nonce'], 'wcml_update_currency_lang')){
+        $nonce = filter_input( INPUT_POST, 'wcml_nonce', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+        if(!$nonce || !wp_verify_nonce($nonce, 'wcml_update_currency_lang')){
             die('Invalid nonce');
         }
         global $woocommerce_wpml;
@@ -354,13 +407,13 @@ class WCML_Multi_Currency_Support{
     }
 
     function update_default_currency(){
-        if(!wp_verify_nonce($_REQUEST['wcml_nonce'], 'wcml_update_default_currency')){
+        $nonce = filter_input( INPUT_POST, 'wcml_nonce', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+        if(!$nonce || !wp_verify_nonce($nonce, 'wcml_update_default_currency')){
             die('Invalid nonce');
         }
         global $woocommerce_wpml;
         $woocommerce_wpml->settings['default_currencies'][$_POST['lang']] = $_POST['code'];
         $woocommerce_wpml->update_settings();
-        
         
         exit;
     }
@@ -600,6 +653,8 @@ class WCML_Multi_Currency_Support{
     }
     
     function load_inline_js(){
+
+        $wcml_switch_currency_nonce = wp_create_nonce('switch_currency');
         
         wc_enqueue_js( "
             jQuery('.wcml_currency_switcher').on('change', function(){                   
@@ -617,18 +672,22 @@ class WCML_Multi_Currency_Support{
                     jQuery('.wcml_currency_switcher').attr('disabled', 'disabled');
                     jQuery('.wcml_currency_switcher').after();
                     ajax_loader.insertAfter(jQuery('.wcml_currency_switcher'));
-                    var data = {action: 'wcml_switch_currency', currency: currency}
                     jQuery.ajax({
                         type : 'post',
                         url : woocommerce_params.ajax_url,
                         data : {
                             action: 'wcml_switch_currency',
-                            currency : currency
+                            currency : currency,
+                            wcml_nonce: '".$wcml_switch_currency_nonce."'
                         },
                         success: function(response) {
-                            jQuery('.wcml_currency_switcher').removeAttr('disabled');
-                        ajax_loader.remove();
-                            window.location = window.location.href;
+                            if(typeof response.error !== 'undefined'){
+                                alert(response.error);
+                            }else{
+                                jQuery('.wcml_currency_switcher').removeAttr('disabled');
+                                ajax_loader.remove();
+                                window.location = window.location.href;
+                            }
                         }
                     });
             }
@@ -713,7 +772,6 @@ class WCML_Multi_Currency_Support{
                             }else{
                                 $variation_fields[$k][$j] = apply_filters('wcml_raw_price_amount', $amount );   // automatic conversion
                             }
-                            
                             
                         }
                         
@@ -820,20 +878,20 @@ class WCML_Multi_Currency_Support{
                 $variation_ids = $wpdb->get_col($wpdb->prepare("SELECT ID FROM {$wpdb->posts} WHERE post_parent = %d", $product_id));
                 
                 // variations with custom prices
-                $res = $wpdb->get_results("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(" . join(',', $variation_ids) . ") AND meta_key='_wcml_custom_prices_status'");
+                $res = $wpdb->get_results($wpdb->prepare("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(%s) AND meta_key='_wcml_custom_prices_status'",join(',', $variation_ids)));
                 foreach($res as $row){
                     $custom_prices_enabled[$row->post_id] = $row->meta_value;
                 }
                 
                 // REGULAR PRICES
                 // get custom prices
-                $res = $wpdb->get_results("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(" . join(',', $variation_ids) . ") AND meta_key='_regular_price_" . $currency . "'");
+                $res = $wpdb->get_results($wpdb->prepare("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(%s) AND meta_key='_regular_price_" . $currency . "'",join(',', $variation_ids)));
                 foreach($res as $row){
                     $regular_prices[$row->post_id] = $row->meta_value;
                 }
                 
                 // get default prices (default currency)
-                $res = $wpdb->get_results("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(" . join(',', $variation_ids) . ") AND meta_key='_regular_price'");
+                $res = $wpdb->get_results($wpdb->prepare("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(%s) AND meta_key='_regular_price'",join(',', $variation_ids)));
                 foreach($res as $row){
                     $default_regular_prices[$row->post_id] = $row->meta_value;
                 }
@@ -847,13 +905,13 @@ class WCML_Multi_Currency_Support{
                 
                 // SALE PRICES
                 // get custom prices
-                $res = $wpdb->get_results("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(" . join(',', $variation_ids) . ") AND meta_key='_sale_price_" . $currency . "'");
+                $res = $wpdb->get_results($wpdb->prepare("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(%s) AND meta_key=%s",join(',', $variation_ids),'_sale_price_'.$currency));
                 foreach($res as $row){
                     $custom_sale_prices[$row->post_id] = $row->meta_value;
                 }
                 
                 // get default prices (default currency)
-                $res = $wpdb->get_results("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(" . join(',', $variation_ids) . ") AND meta_key='_sale_price' AND meta_value <> ''");
+                $res = $wpdb->get_results($wpdb->prepare("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(%s) AND meta_key='_sale_price' AND meta_value <> ''",join(',', $variation_ids)));
                 foreach($res as $row){
                     $default_sale_prices[$row->post_id] = $row->meta_value;
                 }
@@ -868,13 +926,13 @@ class WCML_Multi_Currency_Support{
                 
                 // PRICES
                 // get custom prices
-                $res = $wpdb->get_results("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(" . join(',', $variation_ids) . ") AND meta_key='_price_" . $currency . "'");
+                $res = $wpdb->get_results($wpdb->prepare("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(%s) AND meta_key=%s",join(',', $variation_ids),'_price_'.$currency));
                 foreach($res as $row){
                     $custom_prices_prices[$row->post_id] = $row->meta_value;
                 }
                 
                 // get default prices (default currency)
-                $res = $wpdb->get_results("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(" . join(',', $variation_ids) . ") AND meta_key='_price'");
+                $res = $wpdb->get_results($wpdb->prepare("SELECT post_id, meta_value FROM {$wpdb->postmeta} WHERE post_id IN(%s) AND meta_key='_price'",join(',', $variation_ids)));
                 foreach($res as $row){
                     $default_prices[$row->post_id] = $row->meta_value;
                 }
@@ -1026,7 +1084,7 @@ class WCML_Multi_Currency_Support{
         $active_languages     = $sitepress->get_active_languages();
         
         if(isset($_POST['action']) && $_POST['action'] == 'wcml_switch_currency' && !empty($_POST['currency'])){
-           $this->client_currency = $_POST['currency'];
+           $this->client_currency = filter_input( INPUT_POST, 'currency', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
         }
 
         if( is_null($this->client_currency) && isset($default_currencies[$current_language]) && $default_currencies[$current_language] && !empty($woocommerce->session) && $current_language != $woocommerce->session->get('client_currency_language') ){
@@ -1083,7 +1141,15 @@ class WCML_Multi_Currency_Support{
     }
     
     function switch_currency(){
-        $this->set_client_currency($_POST['currency']);
+        $nonce = filter_input( INPUT_POST, 'wcml_nonce', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+        if(!$nonce || !wp_verify_nonce($nonce, 'switch_currency')){
+            echo json_encode(array('error' => __('Invalid nonce', 'wpml-wcml')));
+            die();
+        }
+
+        $currency = filter_input( INPUT_POST, 'currency', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+
+        $this->set_client_currency($currency);
         
         // force set user cookie when user is not logged in        
         global $woocommerce, $current_user;
@@ -1091,7 +1157,7 @@ class WCML_Multi_Currency_Support{
             $woocommerce->session->set_customer_session_cookie(true);    
         }
         
-        do_action('wcml_switch_currency', $_POST['currency']);
+        do_action('wcml_switch_currency', $currency );
         
         exit;
         
@@ -1151,7 +1217,12 @@ class WCML_Multi_Currency_Support{
     }
 
     function legacy_update_custom_rates(){
-        
+
+        $nonce = filter_input( INPUT_POST, 'wcml_nonce', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+        if(!$nonce || !wp_verify_nonce($nonce, 'legacy_update_custom_rates')){
+            die('Invalid nonce');
+        }
+
         foreach($_POST['posts'] as $post_id => $rates){
             
             update_post_meta($post_id, '_custom_conversion_rate', $rates);
@@ -1164,7 +1235,13 @@ class WCML_Multi_Currency_Support{
     }
     
     function legacy_remove_custom_rates(){
-        
+
+        $nonce = filter_input( INPUT_POST, 'wcml_nonce', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+        if(!$nonce || !wp_verify_nonce($nonce, 'legacy_remove_custom_rates')){
+            echo json_encode(array('error' => __('Invalid nonce', 'wpml-wcml')));
+            die();
+        }
+
         delete_post_meta($_POST['post_id'], '_custom_conversion_rate');
         echo json_encode(array());
